Clarify semantics of append parameter
The command-line help already states that the append parameter has
no effect when there is no existing dataset (i.e. in this case the
output zarr will be created whether append is true or false).
This commit also documents this behaviour in the template
configuration file and the docstring for Converter.__init__.