Do pre-append checks for dry runs as well
writer._check_append_allowed is now called even when doing a dry run.
I've also added an extra check to _check_append_allowed to confirm
that the output Zarr exists before opening it, because in a dry run
writer._append_dataset may be called with no actual dataset to append
to.